requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py @ git+https://github.com/rapptz/discord.py@8fd1fd805ae9fcb540b56d34152087a4979b57c5
frozenlist==1.4.1
g4f==0.3.0.8
idna==3.7
multidict==6.0.5
nest-asyncio==1.6.0
pycryptodome==3.20.0
requests==2.31.0
urllib3==2.2.1
yarl==1.9.4
